feat: global smooth scroll via CSS instead of JS in one nav component (#170)
This PR factors out custom event handlers and a smooth scroll behavior
that targeted only the on-page Home navigation links.

Instead of hijacking the click events for those links and using
JavaScript to apply the smooth scroll behavior, I'm using CSS to apply
that behavior to every page's **html** element. If this turns out to be
too much we can narrow the scope down to specific pages and/or elements.

I'm also adding a new home page left navigation menu item "Home" that
uses the anchor element **#top** to navigate users back to the top of
the page.