content: add new blog post on the history of this site (#189)
This PR adds a new blog summarizing the evolution of my website. Well –
technically there is a step from 2018-2019 I'm missing and still trying
to track down.

I added one hack – in the CSS, I set a `max-width` for the blog post
images because I was struggling to use the MDX renderer to override the
image component styles.