chore: migrate from CircleCI to GitHub Action (#300)
This pull request replaces the CircleCI configuration with a GitHub
Actions workflow for continuous integration. The new workflow maintains
similar functionality while aligning with GitHub Actions' structure and
capabilities.

### Migration from CircleCI to GitHub Actions:

*
[`.circleci/config.yml`](diffhunk://#diff-78a8a19706dbd2a4425dd72bdab0502ed7a2cef16365ab7030a5a0588927bf47L1-L46):
Removed the CircleCI configuration file, including jobs for linting,
testing, and uploading coverage reports.

*
[`.github/workflows/ci.yml`](diffhunk://#diff-b803fcb7f17ed9235f1e5cb1fcd2f5d3b2838429d4368ae4c57ce4436577f03fR1-R81):
Added a new GitHub Actions workflow with jobs for linting, testing,
uploading coverage to Codecov, and an optional build step. The workflow
is triggered on pushes and pull requests to the `main` branch.