fix: recently read book scrolling in Safari (#322)
This PR fixes 1) fixes a bug in Safari where clicking a book in the book
explorer would scroll to the top of the page. It also 2) adds a new
behavior to scroll to the book explorer when opened with a **bookId**
query param.

### Screenshot

A demonstration of the bug(s) in Safari.


https://github.com/user-attachments/assets/c08e1786-a85f-4e60-af05-e70dce12b173

A demonstration of the issue fixed via the deploy preview link.


https://github.com/user-attachments/assets/ae4d888b-ada7-497f-911f-e67b7968f713