content: clean up the latest blog post text (#338)
This PR does a pass at cleaning up the text on my latest blog post. It
also adds a screenshot of the app described in that post.

## AI summary

This pull request includes updates to the blog post about the piano
repertoire and a version bump in the `package.json` file. The blog post
revisions improve readability and enhance user engagement, while the
version update reflects these changes.

### Blog post updates:
*
[`www.chrisvogt.me/content/blog/2025-01-07-my-piano-repertoire.mdx`](diffhunk://#diff-07ff2c8353f84b24100973541f673a30b010b2bb51b5d8d081c61c3902dffc3eL23-R33):
Revised the blog post to improve clarity and engagement, including
rephrased sentences for better readability, added a screenshot of the
repertoire app, and updated links to make them more user-friendly.

### Version update:
*
[`www.chrisvogt.me/package.json`](diffhunk://#diff-345f7dcb2c5893ce5467041532f4bf4892e2f6b8c97661561f45faba61d7afe9L4-R4):
Bumped the version number from `1.5.1` to `1.5.2` to reflect the blog
post updates.