hotfix: busts cache for Flowers for Algernon (#359)
I find the API data returned for the _Flowers for Algernon_ book
description rude and offensive. As a quick fix, I updated the data in
the API and patched in new content. This adds a new query param to the
data source URL to force the cache to bust on the client.